pyproject.toml
[project]


name = "athena_client"
version = "0.1.0"
description = "Athena Client Library"
readme = "README.md"
authors = [{ name = "Resolver Global", email = "opensource@kroll.com" }]
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.10.0",
    "brotli>=1.1.0",
    "grpcio-tools>=1.74.0",
    "numpy>=2.2.6",
    "pillow>=11.3.0",
]
[dependency-groups]
dev = [
    "mypy-protobuf>=3.6.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
    "types-protobuf>=6.30.2.20250703",
]

[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 80
exclude = ["src/athena_client/generated/*"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D213", "D211", "D203", "S101"]

[tool.ruff.lint.per-file-ignores]
# Ignore doc lint rules in tests.
"tests/**" = ["D"]

[tool.pyright]
exclude = ["src/athena_client/generated/*", ".venv"]
venvPath = "."
venv = ".venv"
